# Conflicts:
#	README.md
#	mkdocs.yml
#	src/OrchardCore.Build/Dependencies.props
#	src/OrchardCore.Build/OrchardCore.Commons.props
#	src/OrchardCore.Modules/OrchardCore.ContentLocalization/Sitemaps/LocalizedContentItemsQueryProvider.cs
#	src/OrchardCore.Modules/OrchardCore.Contents/Sitemaps/DefaultContentItemsQueryProvider.cs
#	src/OrchardCore.Modules/OrchardCore.XmlRpc/Services/XmlRpcWriter.cs
#	src/OrchardCore/OrchardCore.Search.Elasticsearch.Core/Services/ElasticIndexManager.cs
#	src/OrchardCore/OrchardCore.Search.Elasticsearch.Core/Services/ElasticQueryService.cs
#	src/docs/README.md
#	src/docs/getting-started/templates/README.md

+++ src/OrchardCore/OrchardCore.FileStorage.AmazonS3/AwsFileStorage.cs
@@ -1,18 +1,10 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
 using System.Net;
-using System.Threading.Tasks;
 using Amazon.S3;
 using Amazon.S3.Model;
 using Amazon.S3.Transfer;
 using OrchardCore.Modules;
 
 namespace OrchardCore.FileStorage.AmazonS3;
-
-using static String;
-
 /// <summary>
 /// Provides an <see cref="IFileStore"/> implementation that targets an underlying Amazon S3 Bucket account.
 /// </summary>
@@ -29,7 +21,7 @@ public class AwsFileStore : IFileStore
         _options = options;
         _amazonS3Client = amazonS3Client;
 
-        if (!IsNullOrEmpty(_options.BasePath))
+        if (!string.IsNullOrEmpty(_options.BasePath))
         {
             _basePrefix = NormalizePrefix(_options.BasePath);
         }
@@ -56,7 +48,7 @@ public class AwsFileStore : IFileStore
 
     public async Task<IFileStoreEntry> GetDirectoryInfoAsync(string path)
     {
-        if (IsNullOrEmpty(path))
+        if (string.IsNullOrEmpty(path))
         {
             return new AwsDirectory(path, _clock.UtcNow);
         }
@@ -69,7 +61,7 @@ public class AwsFileStore : IFileStore
             FetchOwner = false
         });
 
-        return awsDirectory.S3Objects.Any() ? new AwsDirectory(path, _clock.UtcNow) : null;
+        return awsDirectory.S3Objects.Count > 0 ? new AwsDirectory(path, _clock.UtcNow) : null;
     }
 
     public async IAsyncEnumerable<IFileStoreEntry> GetDirectoryContentAsync(string path = null,
@@ -87,9 +79,9 @@ public class AwsFileStore : IFileStore
         {
             var itemName = Path.GetFileName(WebUtility.UrlDecode(file.Key));
 
-            if (includeSubDirectories || !IsNullOrEmpty(itemName))
+            if (includeSubDirectories || !string.IsNullOrEmpty(itemName))
             {
-                if (IsNullOrEmpty(path))
+                if (string.IsNullOrEmpty(path))
                 {
                     path = "/";
                 }
@@ -101,7 +93,7 @@ public class AwsFileStore : IFileStore
         foreach (var awsFolderPath in listObjectsResponse.CommonPrefixes)
         {
             var folderPath = awsFolderPath;
-            if (!IsNullOrEmpty(_basePrefix))
+            if (!string.IsNullOrEmpty(_basePrefix))
             {
                 folderPath = folderPath[(_basePrefix.Length - 1)..];
             }
@@ -149,7 +141,7 @@ public class AwsFileStore : IFileStore
 
     public async Task<bool> TryDeleteDirectoryAsync(string path)
     {
-        if (IsNullOrWhiteSpace(path))
+        if (string.IsNullOrWhiteSpace(path))
         {
             throw new FileStoreException("Cannot delete root directory.");
         }
@@ -160,16 +152,20 @@ public class AwsFileStore : IFileStore
             Prefix = NormalizePrefix(this.Combine(_basePrefix, path))
         });
 
-        var deleteObjectsRequest = new DeleteObjectsRequest
+        if (listObjectsResponse.S3Objects.Count > 0)
         {
-            BucketName = _options.BucketName,
-            Objects = listObjectsResponse.S3Objects
-                .Select(metadata => new KeyVersion { Key = metadata.Key }).ToList()
-        };
+            var deleteObjectsRequest = new DeleteObjectsRequest
+            {
+                BucketName = _options.BucketName,
+                Objects = listObjectsResponse.S3Objects
+                    .Select(metadata => new KeyVersion { Key = metadata.Key }).ToList()
+            };
 
-        var response = await _amazonS3Client.DeleteObjectsAsync(deleteObjectsRequest);
+            var response = await _amazonS3Client.DeleteObjectsAsync(deleteObjectsRequest);
+            return response.IsSuccessful();
+        }
 
-        return response.IsSuccessful();
+        return listObjectsResponse.IsSuccessful();
     }
 
     public async Task MoveFileAsync(string oldPath, string newPath)
@@ -206,7 +202,7 @@ public class AwsFileStore : IFileStore
                 Prefix = this.Combine(_basePrefix, dstPath)
             });
 
-            if (listObjects.S3Objects.Any())
+            if (listObjects.S3Objects.Count > 0)
             {
                 throw new FileStoreException($"Cannot copy file '{srcPath}' because a file already exists in the new path '{dstPath}'.");
             }
@@ -261,7 +257,7 @@ public class AwsFileStore : IFileStore
                     Prefix = this.Combine(_basePrefix, path)
                 });
 
-                if (listObjects.S3Objects.Any())
+                if (listObjects.S3Objects.Count > 0)
                 {
                     throw new FileStoreException($"Cannot create file '{path}' because it already exists.");
                 }
@@ -292,7 +288,7 @@ public class AwsFileStore : IFileStore
         prefix = prefix.Trim('/') + '/';
         if (prefix.Length == 1)
         {
-            return Empty;
+            return string.Empty;
         }
 
         return prefix;
